requirements.txt
Django==4.1.2
django-ckeditor==6.5.1